requirements.txt
aiohttp~=3.8.4
pillow
playwright~=1.32.1